+++ src/reconstruction/kalman_filter_linear.cpp
@@ -88,8 +88,8 @@ bool KalmanFilterLinear::validateState(const kalman::KalmanState& s, bool xP_onl
     if (!KalmanFilter::validateState(s, xP_only))
         return false;
 
-    if (!xP_only && (s.F.cols() != dim_x_ ||
-                     s.F.rows() != dim_x_))
+    if (!xP_only && (static_cast<size_t>(s.F.cols()) != dim_x_ ||
+                     static_cast<size_t>(s.F.rows()) != dim_x_))
         return false;
 
     return true;
@@ -102,8 +102,8 @@ bool KalmanFilterLinear::validateState(const kalman::BasicKalmanState& s, bool x
     if (!KalmanFilter::validateState(s, xP_only))
         return false;
 
-    if (!xP_only && (s.F.cols() != dim_x_ ||
-                     s.F.rows() != dim_x_))
+    if (!xP_only && (static_cast<size_t>(s.F.cols()) != dim_x_ ||
+                     static_cast<size_t>(s.F.rows()) != dim_x_))
         return false;
 
     return true;
